SDIT-2832 Split sync service into one per entity
